Fixed Runtime Errors
Ran the drone, encountered multiple run time errors
specifically with multi inheritance in TelloDrone.py
where the parent classes constructors were not being
called due to them not implementing super().

Need to visualize what state we are in for debugging.
I was able to takeoff but was unable to do anything
but land -> I do not know if a state change actually
happened.

Write more tests!!!